+++ packages/components-native/src/AtlantisContext/AtlantisContext.tsx
@@ -55,6 +55,13 @@ export interface AtlantisContextProps {
    * @default "en"
    */
   readonly locale: string;
+
+  /**
+   * The `setHeaderHeight` method allows you to set the height of the app header in Atlantis.
+   * Adjusting this height is essential for ensuring the correct positioning and alignment of various elements within the app.
+   * By setting this value accurately, Atlantis can dynamically adjust the layout of its components based on the specified header height.
+   */
+  readonly setHeaderHeight: (height: number) => void;
 }
 
 export const defaultValues: AtlantisContextProps = {
@@ -70,6 +77,9 @@ export const defaultValues: AtlantisContextProps = {
   currencySymbol: DEFAULT_CURRENCY_SYMBOL,
   headerHeight: 0,
   locale: "en",
+  setHeaderHeight: _ => {
+    return;
+  },
 };
 
 export const AtlantisContext = createContext(defaultValues);