+++ packages/components/src/DataList/__tests__/DataList.utils.test.tsx
@@ -4,7 +4,10 @@ import {
   generateHeaderElements,
   generateListItemElements,
   getCompoundComponent,
+  getCompoundComponents,
+  sortSizeProp,
 } from "../DataList.utils";
+import { BREAKPOINTS } from "../DataList.const";
 
 const date = new Date("2001-01-01T00:00:00.000Z");
 
@@ -47,7 +50,7 @@ describe("Datalist utils", () => {
    */
   describe("generateListItemElements", () => {
     it("should generate a Text component for the label key", () => {
-      const elementList = generateListItemElements([{ label: "Hello" }]);
+      const elementList = generateListItemElements([{ id: 1, label: "Hello" }]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].label).toMatchInlineSnapshot(`
@@ -60,6 +63,7 @@ describe("Datalist utils", () => {
     it("should generate a subdued Text component for any random key", () => {
       const elementList = generateListItemElements([
         {
+          id: 1,
           randomKeyThatIsntNormal: "I am a normal text",
         },
       ]);
@@ -75,38 +79,34 @@ describe("Datalist utils", () => {
     });
 
     it("should generate a list of inline label for the tag key", () => {
-      const elementList = generateListItemElements([{ tags: ["uno", "dos"] }]);
+      const elementList = generateListItemElements([
+        { id: 1, tags: ["uno", "dos"] },
+      ]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].tags).toMatchInlineSnapshot(`
-        <div
-          style={
-            {
-              "display": "flex",
-              "gap": 8,
-              "overflow": "hidden",
-            }
+        <DataListTags
+          items={
+            [
+              "uno",
+              "dos",
+            ]
           }
-        >
-          <InlineLabel>
-            uno
-          </InlineLabel>
-          <InlineLabel>
-            dos
-          </InlineLabel>
-        </div>
+        />
       `);
     });
 
     it("should generate the element passed in on any key", () => {
-      const elementList = generateListItemElements([{ element: <div /> }]);
+      const elementList = generateListItemElements([
+        { id: 1, element: <div /> },
+      ]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].element).toMatchInlineSnapshot(`<div />`);
     });
 
     it("should generate the correct element", () => {
-      const elementList = generateListItemElements([{ date }]);
+      const elementList = generateListItemElements([{ id: 1, date }]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].date).toMatchInlineSnapshot(`
@@ -136,6 +136,37 @@ describe("Datalist utils", () => {
     it("should return undefined when the value is empty", () => {
       const headerElements = generateHeaderElements({});
       expect(headerElements).toBeUndefined();
+    });
+  });
+
+  describe("Sort Size Prop", () => {
+    it("sorts correctly", () => {
+      expect(sortSizeProp(["xl", "md", "lg", "sm", "xs"])).toEqual(BREAKPOINTS);
+    });
+  });
+
+  describe("getCompoundComponents", () => {
+    const Component = () => <></>;
+    const ComponentDos = () => <></>;
+    it("should return the components I'm looking for", () => {
+      const children = [
+        <Component key="1" />,
+        <ComponentDos key="2" />,
+        <Component key="3" />,
+      ];
+      const components = getCompoundComponents(children, Component);
+
+      expect(components[0]?.type).toBe(Component);
+      expect(components[0]?.type).not.toBe(ComponentDos);
+      expect(components[1]?.type).toBe(Component);
+      expect(components[1]?.type).not.toBe(ComponentDos);
+    });
+
+    it("should return an empty list if the component i'm looking for doesn't exist", () => {
+      const children = [<ComponentDos key="2" />];
+      const component = getCompoundComponents(children, Component);
+
+      expect(component).toHaveLength(0);
     });
   });
 });