+++ packages/components/src/DataList/components/DataListLoadingState/DataListLoadingState.css.d.ts
@@ -1,0 +1,8 @@
+declare const styles: {
+  readonly "loadingItem": string;
+  readonly "mobileLoadingState": string;
+  readonly "firstMobileGlimmer": string;
+  readonly "thirdMobileGlimmer": string;
+};
+export = styles;
+